TESTING: performance optimization for monitors - use float parsing instead int parsing for frequencies to allow frequencies < 1

+++ jumper_extension/monitor/benchmark/run_benchmark.py
@@ -655,7 +655,7 @@ def main():
         else list(BACKENDS.keys())
     )
     frequencies = (
-        [int(f) for f in args.frequencies.split(",")]
+        [float(f) for f in args.frequencies.split(",")]
         if args.frequencies else FREQUENCIES
     )
